# gbfs_visualizer_server/openapi_server/models/gbfs_free_bike_status.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class GbfsFreeBikeStatus(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, gbfs_system_id=None, url=None, bikes=None):  # noqa: E501
        """GbfsFreeBikeStatus - a model defined in OpenAPI

        :param gbfs_system_id: The gbfs_system_id of this GbfsFreeBikeStatus.  # noqa: E501
        :type gbfs_system_id: str
        :param url: The url of this GbfsFreeBikeStatus.  # noqa: E501
        :type url: str
        :param bikes: The bikes of this GbfsFreeBikeStatus.  # noqa: E501
        :type bikes: List[object]
        """
        self.openapi_types = {
            'gbfs_system_id': str,
            'url': str,
            'bikes': List[object]
        }

        self.attribute_map = {
            'gbfs_system_id': 'gbfs_system_id',
            'url': 'url',
            'bikes': 'bikes'
        }

        self._gbfs_system_id = gbfs_system_id
        self._url = url
        self._bikes = bikes

    @classmethod
    def from_dict(cls, dikt) -> 'GbfsFreeBikeStatus':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The gbfs_free_bike_status of this GbfsFreeBikeStatus.  # noqa: E501
        :rtype: GbfsFreeBikeStatus
        """
        return util.deserialize_model(dikt, cls)

    @property
    def gbfs_system_id(self):
        """Gets the gbfs_system_id of this GbfsFreeBikeStatus.

        ID of the related GBFS System  # noqa: E501

        :return: The gbfs_system_id of this GbfsFreeBikeStatus.
        :rtype: str
        """
        return self._gbfs_system_id

    @gbfs_system_id.setter
    def gbfs_system_id(self, gbfs_system_id):
        """Sets the gbfs_system_id of this GbfsFreeBikeStatus.

        ID of the related GBFS System  # noqa: E501

        :param gbfs_system_id: The gbfs_system_id of this GbfsFreeBikeStatus.
        :type gbfs_system_id: str
        """
        if gbfs_system_id is not None and len(gbfs_system_id) > 256:
            raise ValueError("Invalid value for `gbfs_system_id`, length must be less than or equal to `256`")  # noqa: E501

        self._gbfs_system_id = gbfs_system_id

    @property
    def url(self):
        """Gets the url of this GbfsFreeBikeStatus.

        URL to the station_status.json file of the GBFS System, containing all the information related with the correspondent stations status of the system  # noqa: E501

        :return: The url of this GbfsFreeBikeStatus.
        :rtype: str
        """
        return self._url

    @url.setter
    def url(self, url):
        """Sets the url of this GbfsFreeBikeStatus.

        URL to the station_status.json file of the GBFS System, containing all the information related with the correspondent stations status of the system  # noqa: E501

        :param url: The url of this GbfsFreeBikeStatus.
        :type url: str
        """

        self._url = url

    @property
    def bikes(self):
        """Gets the bikes of this GbfsFreeBikeStatus.


        :return: The bikes of this GbfsFreeBikeStatus.
        :rtype: List[object]
        """
        return self._bikes

    @bikes.setter
    def bikes(self, bikes):
        """Sets the bikes of this GbfsFreeBikeStatus.


        :param bikes: The bikes of this GbfsFreeBikeStatus.
        :type bikes: List[object]
        """

        self._bikes = bikes