# gbfs_visualizer_server/openapi_server/models/gbfs_station_information.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class GbfsStationInformation(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, gbfs_system_id=None, url=None, stations=None):  # noqa: E501
        """GbfsStationInformation - a model defined in OpenAPI

        :param gbfs_system_id: The gbfs_system_id of this GbfsStationInformation.  # noqa: E501
        :type gbfs_system_id: str
        :param url: The url of this GbfsStationInformation.  # noqa: E501
        :type url: str
        :param stations: The stations of this GbfsStationInformation.  # noqa: E501
        :type stations: List[object]
        """
        self.openapi_types = {
            'gbfs_system_id': str,
            'url': str,
            'stations': List[object]
        }

        self.attribute_map = {
            'gbfs_system_id': 'gbfs_system_id',
            'url': 'url',
            'stations': 'stations'
        }

        self._gbfs_system_id = gbfs_system_id
        self._url = url
        self._stations = stations

    @classmethod
    def from_dict(cls, dikt) -> 'GbfsStationInformation':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The gbfs_station_information of this GbfsStationInformation.  # noqa: E501
        :rtype: GbfsStationInformation
        """
        return util.deserialize_model(dikt, cls)

    @property
    def gbfs_system_id(self):
        """Gets the gbfs_system_id of this GbfsStationInformation.


        :return: The gbfs_system_id of this GbfsStationInformation.
        :rtype: str
        """
        return self._gbfs_system_id

    @gbfs_system_id.setter
    def gbfs_system_id(self, gbfs_system_id):
        """Sets the gbfs_system_id of this GbfsStationInformation.


        :param gbfs_system_id: The gbfs_system_id of this GbfsStationInformation.
        :type gbfs_system_id: str
        """

        self._gbfs_system_id = gbfs_system_id

    @property
    def url(self):
        """Gets the url of this GbfsStationInformation.

        URL to the station_information.json file of the GBFS System, containing all the information related with the correspondent stations of the system  # noqa: E501

        :return: The url of this GbfsStationInformation.
        :rtype: str
        """
        return self._url

    @url.setter
    def url(self, url):
        """Sets the url of this GbfsStationInformation.

        URL to the station_information.json file of the GBFS System, containing all the information related with the correspondent stations of the system  # noqa: E501

        :param url: The url of this GbfsStationInformation.
        :type url: str
        """

        self._url = url

    @property
    def stations(self):
        """Gets the stations of this GbfsStationInformation.


        :return: The stations of this GbfsStationInformation.
        :rtype: List[object]
        """
        return self._stations

    @stations.setter
    def stations(self, stations):
        """Sets the stations of this GbfsStationInformation.


        :param stations: The stations of this GbfsStationInformation.
        :type stations: List[object]
        """

        self._stations = stations