# gbfs_visualizer_server/openapi_server/models/gbfs_station_status.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class GbfsStationStatus(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, gbfs_system_id=None, url=None, stations=None):  # noqa: E501
        """GbfsStationStatus - a model defined in OpenAPI

        :param gbfs_system_id: The gbfs_system_id of this GbfsStationStatus.  # noqa: E501
        :type gbfs_system_id: str
        :param url: The url of this GbfsStationStatus.  # noqa: E501
        :type url: str
        :param stations: The stations of this GbfsStationStatus.  # noqa: E501
        :type stations: List[object]
        """
        self.openapi_types = {
            'gbfs_system_id': str,
            'url': str,
            'stations': List[object]
        }

        self.attribute_map = {
            'gbfs_system_id': 'gbfs_system_id',
            'url': 'url',
            'stations': 'stations'
        }

        self._gbfs_system_id = gbfs_system_id
        self._url = url
        self._stations = stations

    @classmethod
    def from_dict(cls, dikt) -> 'GbfsStationStatus':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The gbfs_station_status of this GbfsStationStatus.  # noqa: E501
        :rtype: GbfsStationStatus
        """
        return util.deserialize_model(dikt, cls)

    @property
    def gbfs_system_id(self):
        """Gets the gbfs_system_id of this GbfsStationStatus.

        ID of the related GBFS System  # noqa: E501

        :return: The gbfs_system_id of this GbfsStationStatus.
        :rtype: str
        """
        return self._gbfs_system_id

    @gbfs_system_id.setter
    def gbfs_system_id(self, gbfs_system_id):
        """Sets the gbfs_system_id of this GbfsStationStatus.

        ID of the related GBFS System  # noqa: E501

        :param gbfs_system_id: The gbfs_system_id of this GbfsStationStatus.
        :type gbfs_system_id: str
        """
        if gbfs_system_id is not None and len(gbfs_system_id) > 256:
            raise ValueError("Invalid value for `gbfs_system_id`, length must be less than or equal to `256`")  # noqa: E501

        self._gbfs_system_id = gbfs_system_id

    @property
    def url(self):
        """Gets the url of this GbfsStationStatus.

        URL to the station_status.json file of the GBFS System, containing all the information related with the correspondent stations status of the system  # noqa: E501

        :return: The url of this GbfsStationStatus.
        :rtype: str
        """
        return self._url

    @url.setter
    def url(self, url):
        """Sets the url of this GbfsStationStatus.

        URL to the station_status.json file of the GBFS System, containing all the information related with the correspondent stations status of the system  # noqa: E501

        :param url: The url of this GbfsStationStatus.
        :type url: str
        """

        self._url = url

    @property
    def stations(self):
        """Gets the stations of this GbfsStationStatus.


        :return: The stations of this GbfsStationStatus.
        :rtype: List[object]
        """
        return self._stations

    @stations.setter
    def stations(self, stations):
        """Sets the stations of this GbfsStationStatus.


        :param stations: The stations of this GbfsStationStatus.
        :type stations: List[object]
        """

        self._stations = stations